requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.26.0
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
slack-sdk==3.33.1
